refactor: move common to theme folder
